+++ src/types/KeyHandler.d.ts
@@ -1,0 +1,13 @@
+export default interface KeyHandlerProps {
+    componentName: string ,
+    onEnter?: () => void;
+    onBack?: () => void;
+    onArrowUp?: () => void;
+    onArrowDown?: () => void;
+    onArrowLeft?: () => void;
+    onArrowRight?: () => void;
+    onRed?: () => void;
+    onGreen?: () => void;
+    onYellow?: () => void;
+    onBlue?: () => void;
+  }